requirements.txt
Pillow
numpy
picamera2
smbus